+++ CHANGES.rst
@@ -16,6 +16,9 @@ Changelog
   is now accessible to `advisers` of the historized data advice and
   proposingGroup members.
   [gbastien]
+- Make sure `data_changes` history does not use `highlight_last_comment` or
+  it drastically slows down item view when used.
+  [gbastien]
 
 4.2.5 (2023-09-12)
 ------------------